+++ src-docs/src/views/tables/selection/selection_section.js
@@ -22,8 +22,12 @@ export const section = {
   text: (
     <p>
       The following example shows how to configure selection via the{' '}
-      <EuiCode>selection</EuiCode>
-      property.
+      <EuiCode>selection</EuiCode> property.
+      <br />
+      This is important to provide an <EuiCode>ariaLabel</EuiCode> property of{' '}
+      selection table items to supply a valuable screen reader user experience.
+      <br />
+      Default value is <EuiCode>Select this row</EuiCode>.
     </p>
   ),
   components: { EuiBasicTable },